Relate blink detection with change point finding
I found that when a blink is made, the EAR changes dramatically and
rapidly, which can be generalized to a "change point finding" problem.

A window-based approach is introduced in this commit with the
standard deviation as its cost function.
Refer to https://www.iese.fraunhofer.de/blog/change-point-detection/
for more information.